+++ client/src/views/SprintDashboard/rollupBoard/boardLayout.test.ts
@@ -74,6 +74,7 @@ function buildItem(input: BuildItemInput): RollupBoardItem {
     checklistCompletion: null,
     checklistItems: [],
     isFlagged: false,
+    impedimentReasons: [],
   };
 }